Restructuring & Renaming for making it future proof
