+++ main.go
@@ -4,7 +4,7 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/siesgstarena/epicentre/config"
 	routes "github.com/siesgstarena/epicentre/router"
-	"github.com/siesgstarena/epicentre/services/logger"
+	"github.com/siesgstarena/epicentre/logger"
 	"github.com/siesgstarena/epicentre/services/mongo"
 )
