+++ model/rule.go
@@ -2,8 +2,8 @@ package model
 
 import "go.mongodb.org/mongo-driver/bson/primitive"
 
-//Rules Type exported for use in API
-type Rules struct {
+//Rule Type exported for use in API
+type Rule struct {
 	ID     					primitive.ObjectID 	`bson:"_id,omitempty"`
 	UserID					primitive.ObjectID 	`bson:"userid,omitempty"`
 	ProjectID   			primitive.ObjectID	`bson:"projectid,omitempty"`
