+++ model/user.go
@@ -2,8 +2,8 @@ package model
 
 import "go.mongodb.org/mongo-driver/bson/primitive"
 
-//Users Type exported for use in API
-type Users struct {
+//User Type exported for use in API
+type User struct {
 	ID     		primitive.ObjectID 	`bson:"_id,omitempty"`
 	Name  		string             	`bson:"name,omitempty"`
 	Email 		string             	`bson:"email,omitempty"`
