+++ services/mongo/mongo.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 	"time"
-
 	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/mongo/options"
 	"go.mongodb.org/mongo-driver/mongo/readpref"
